% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_functions.R
\name{check_pheno}
\alias{check_pheno}
\title{Check phenotype file}
\usage{
check_pheno(dsfile, ddfile = NULL, na_vals = c("NA", "N/A", "na", "n/a"),
  subj_exp = NULL, subjectID_col = "SUBJECT_ID")
}
\arguments{
\item{dsfile}{Path to the data file on disk}

\item{ddfile}{Path to the data dictionary file on disk}

\item{na_vals}{Vector of strings that should be read in as NA/missing in data file (see details of \code{.read_ds_file})}

\item{subj_exp}{Vector of expected subject IDs}

\item{subjectID_col}{Column name for subject-level ID}
}
\value{
pheno_report, a list of the following issues (when present):
\item{flag_nonuniq_subjID}{TRUE when subject ID column is not unique, which would require definition of UNIQUEKEY columns in the corresponding data dictionary}
\item{dd_errors}{Differences in fields between data file and data dictionary}
\item{extra_subjects}{Subjects in data file missing from \code{ssm_exp}}
\item{missing_subjects}{Subjects in \code{ssm_exp} missing from data file}
}
\description{
Check contents of a phenotype file for dbGaP posting
}
\details{
Because of the variability of phenotype file contents, the only required column checked here is the subject-level ID. Note dbGaP requests variables (1) described in the study description and/or study config; (2) affection status, if not already included in the subject consent file; (3) sex; and (4) race/ethnicity/ancestry/heritage.

If a data dictionary is provided (\code{ddfile != NULL}), additionally checks 
correspondence between column names in data file and entries in data dictionary.
Data dictionary files can be Excel (.xls, .xlsx) or tab-delimited .txt.
}
