% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_functions.R
\name{.count_hdr_lines}
\alias{.count_hdr_lines}
\title{Count the header lines in a data file}
\usage{
.count_hdr_lines(filename, colname = NA)
}
\arguments{
\item{filename}{The path to the data file on disk}

\item{colname}{One or more expected column names}
}
\value{
the number of header lines in the file
}
\description{
Count the header lines in a data file
}
\details{
Header lines are considered to start with # or to be a blank line.
Optionally, providing a \code{colname} argument, a character vector
of one or more colunm names,  will consider any rows before
that containing the specified column name(s) to be a header row.
}
