% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_functions.R
\name{read_dd_file}
\alias{read_dd_file}
\title{Read data dictionary file}
\usage{
read_dd_file(filename, remove_empty_row = TRUE,
  remove_empty_col = FALSE)
}
\arguments{
\item{filename}{The path to the file on disk}

\item{remove_empty_row}{Logical of whether to exclude empty (i.e. all missing values) rows. Defaults to TRUE}

\item{remove_empty_col}{Logical of whether to exclude empty (i.e. all missing values) rowcolumns. Defaults to FALSE}
}
\value{
A data frame from the file
}
\description{
Read data dictionary file
}
\details{
Expects (tab-delimited) .txt, .xlsx, or .xml file.
For .txt and .xlsx files, dbGaP data dictionary files should have column headers as the first row. If the input violates this, e.g. additional header rows are present, a warning is returned but the file is still read in.
.xml file types should be those provided by dbGaP.
Only a subset ofpossible child nodes of a variable node are processed:
\code{type}
\code{unit}
\code{logical_min}
\code{logical_max}
These names are converted to the names expected in a user-submitted data dictionary.
Finally, if any variables are flagged as "unique keys", a column is added to the output data frame and populated correctly (with X's).
Otherwise, the "UNIQUE_KEY" column does not exist in the output.
}
