% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_functions.R
\name{write_dbgap}
\alias{write_dbgap}
\title{Write dbGaP files}
\usage{
write_dbgap(x, file = "", study_name = NULL, phs = NULL, DD = FALSE,
  dstype = "", generate_fn = FALSE)
}
\arguments{
\item{x}{Object to write out}

\item{file}{Filename. Can leave unspecified if \code{generate_fn=TRUE}}

\item{study_name}{Study name}

\item{phs}{dbGaP PHS accession number}

\item{DD}{logical flag, set to TRUE for DD file, FALSE for DS}

\item{dstype}{Type of file, one of "pheno","ped","sattr","ssm","subj." Can be left as default/NULL if \code{generate_fn=FALSE}}

\item{generate_fn}{logical flag on whether to have filename generated for you}
}
\value{
Write file to disk.
}
\description{
Wrapper to \code{write.table} function to output dbGaP dataset (DS) and data dictionary (DD) files.
}
\details{
If \code{generate_fn=FALSE}, the argument to \code{file} will be used as the file name. File paths will be followed; if none given, will write to current working directory. If \code{generate_fn=TRUE}, one of either \code{study_name} or \code{phs} must be provided. If \code{generate_fn=TRUE}, fill will always be written to current working directory.
}
